all: add rhp package
