+++ rhp/dialer.go
@@ -1,0 +1,47 @@
+package rhp
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"go.sia.tech/core/types"
+	"go.sia.tech/coreutils/rhp/v4"
+	"go.sia.tech/coreutils/rhp/v4/siamux"
+	"go.uber.org/zap"
+)
+
+const dialTimeout = 10 * time.Second
+
+type siamuxDialer struct {
+	cm     ChainManager
+	store  RevisionStore
+	signer rhp.FormContractSigner
+	log    *zap.Logger
+}
+
+// NewSiamuxDialer creates a new Dialer that uses the SiaMux protocol to dial a
+// host.
+func NewSiamuxDialer(cm ChainManager, store RevisionStore, signer rhp.FormContractSigner, log *zap.Logger) Dialer {
+	return &siamuxDialer{
+		cm:     cm,
+		store:  store,
+		signer: signer,
+		log:    log,
+	}
+}
+
+// Dial dials the host and returns a Client that can be used to interact with
+// the host. It uses the SiaMux protocol to establish a connection and returns a
+// host client that exposes the RPC methods defined in the RHP.
+func (d *siamuxDialer) Dial(ctx context.Context, hk types.PublicKey, addr string) (*HostClient, error) {
+	ctx, cancel := context.WithTimeout(ctx, dialTimeout)
+	defer cancel()
+
+	client, err := siamux.Dial(ctx, addr, hk)
+	if err != nil {
+		return nil, fmt.Errorf("failed to dial host: %w", err)
+	}
+
+	return NewHostClient(hk, d.cm, client, d.signer, d.store, d.log.With(zap.Stringer("hostKey", hk))), nil
+}
